requirements.txt
openai>=1.40.0
anthropic>=0.34.0
python-dotenv>=1.0.1
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.2.0